Levels 1-10 done HALF WAY THROUGH! ^_^
Can you make sure there aren't any invisible objects in any levels and remove them is there are. Also an export of this for itch.io uploaded with a PC version for those on chrome would be great. Also adding arrow keys...that's something someone asked for and it's an easy thing I assume